+++ Samples/Example/Scripts/TweeningExample.cs
@@ -12,20 +12,46 @@ namespace SLIDDES.Tweening.Samples
         // Start is called before the first frame update
         void Start()
         {
-            Tween.Value(0, 1, 1).OnChange(x =>
+            // Examples commented
+
+            //Tween.Value(0, 1, 1).OnChange(x =>
+            //{
+            //    //Debug.Log(x.Vector3);
+            //}).OnDestroy(x => Debug.LogWarning("This tween got removed :)"));
+
+            //Tween.Move(moveable, Vector3.zero, new Vector3(1, 0, 0), 1).SetDelay(1);
+
+            //Tween.Move(moveable, Vector3.zero, new Vector3(1, 0, 0), 1).SetEase(Easing.easeInOutCubic).SetDelay(2);
+
+            //Tween.MoveLocalY(moveable, 0, 1, 1).SetDelay(3);
+
+            //Tween.Time(1).OnChange(x =>
+            //{
+            //    if(x.TimeNormalized > 0.5f)
+            //    {
+            //        //Destroy(moveable);
+            //    }
+            //}).SetDelay(4);
+
+            //Tween.Value(new Vector3(0, 0, 0), new Vector3(1, 1, 1), 1).OnChange(x =>
+            //{
+            //    Debug.Log(x.Vector3);
+            //});
+
+            //Tween.MoveX(moveable, 0, 1, 1).Loop(2).SetDelay(5);
+
+            Tween.MoveX(moveable, 0, 5, 5).OnChange(x =>
             {
-                Debug.Log(x.values.vector3.x);
-            }).SetEase(TweenType.easeInElastic);
+                if(x.TimeNormalized > 0.5f) x.ForceComplete();
+            });
 
-            Tween.Move(moveable, Vector3.zero, new Vector3(1, 0, 0), 2).SetEase(TweenType.easeInOutBounce).SetDelay(2);
-
-            TweenInfo tweenInfo = Tween.Time(2).OnComplete(x => Debug.Log("Done A"));
-            tweenInfo.Free(); // interrupt Done A
-            tweenInfo = Tween.Time(2).OnComplete(x => Debug.Log("Done B"));
-
-            Tween.Time(10).OnInterval(1, x => Debug.Log($"Interval {x.values.timer}"));
-
-            Tween.Interval(1, x => Debug.Log("1sec interval"));
+            Tween.Time(10).OnInterval(1, x =>
+            {
+                Debug.Log("1sec Interval");
+            }).OnChange(x => 
+            {
+                if(x.TimeNormalized > 0.5f) x.ForceComplete();
+            });
         }
 
         // Update is called once per frame